Move test language server to app module
