+++ app/build.gradle.kts
@@ -87,6 +87,10 @@ dependencies {
     //Kotlin coroutines
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.0")
 
+    //Xml language server
+    implementation("org.eclipse.lemminx:org.eclipse.lemminx:0.17.0")
+
+
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.3")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.4.0")
