--- editor-lsp/src/main/java/io/github/rosemoe/sora/lsp/mock/MockLanguageConnection.java
@@ -1,55 +1,0 @@
-/*
- *    sora-editor - the awesome code editor for Android
- *    https://github.com/Rosemoe/sora-editor
- *    Copyright (C) 2020-2022  Rosemoe
- *
- *     This library is free software; you can redistribute it and/or
- *     modify it under the terms of the GNU Lesser General Public
- *     License as published by the Free Software Foundation; either
- *     version 2.1 of the License, or (at your option) any later version.
- *
- *     This library is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- *     Lesser General Public License for more details.
- *
- *     You should have received a copy of the GNU Lesser General Public
- *     License along with this library; if not, write to the Free Software
- *     Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
- *     USA
- *
- *     Please contact Rosemoe by email 2073412493@qq.com if you need
- *     additional information or have any questions
- */
-package io.github.rosemoe.sora.lsp.mock;
-
-import android.util.Pair;
-
-import org.eclipse.lemminx.XMLServerLauncher;
-import org.eclipse.lsp4j.jsonrpc.Launcher;
-import org.eclipse.lsp4j.launch.LSPLauncher;
-import org.eclipse.lsp4j.services.LanguageClient;
-import org.eclipse.lsp4j.services.LanguageServer;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.PipedInputStream;
-import java.io.PipedOutputStream;
-import java.net.ServerSocket;
-import java.net.Socket;
-
-public class MockLanguageConnection {
-
-
-    public static void createConnect(int port) throws IOException {
-
-
-        ServerSocket socket = new ServerSocket(port);
-
-        Socket socketClient = socket.accept();
-
-        XMLServerLauncher.launch(socketClient.getInputStream(), socketClient.getOutputStream());
-    }
-
-}
